Create new abstract class to handle REST operation
